Presentation
OWASP Top 10 Maturity Levels for Security Champions 

OWASP needs to support the Community much more on the topic of security champions. We have a project proposing a playbook - that is a good start however there are a great number of details requiring attention for a successful programme. The intention of this talk and project is to provide the necessary guidelines on how to manage security champions and their progress in supporting security. Intentionally the talk is called OWASP Top 10 Maturity Categories for Security Champions - as it relates to the well known Top 10 by suggesting 10 main maturity categories of interest. Because all good ideas need software to materialise, we will be open-sourcing as part of OWASP an application to manage security champions - application that will hopefully evolve as this project with receive more attention.